refactor: Bookmark 서비스를 별도로 분리
