+++ src/test/java/com/techfork/domain/activity/entity/BookmarkTest.java
@@ -1,0 +1,28 @@
+package com.techfork.domain.activity.entity;
+
+import com.techfork.domain.post.entity.Post;
+import com.techfork.domain.useraccount.entity.User;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+
+import java.time.LocalDateTime;
+
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.mockito.Mockito.mock;
+
+class BookmarkTest {
+
+    @Test
+    @DisplayName("북마크 생성 시 사용자, 게시글, 북마크 시각을 그대로 보존한다")
+    void create_PreservesBookmarkState() {
+        User user = mock(User.class);
+        Post post = mock(Post.class);
+        LocalDateTime bookmarkedAt = LocalDateTime.of(2026, 5, 5, 10, 30, 0);
+
+        Bookmark bookmark = Bookmark.create(user, post, bookmarkedAt);
+
+        assertThat(bookmark.getUser()).isSameAs(user);
+        assertThat(bookmark.getPost()).isSameAs(post);
+        assertThat(bookmark.getBookmarkedAt()).isEqualTo(bookmarkedAt);
+    }
+}
